[WIP] Change directory composition & add test code.
