+++ src/yaml_evaluator/go.mod
@@ -1,0 +1,10 @@
+module github.com/at0x0ft/cod2e2/yaml_evaluator
+
+go 1.19
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.8.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
